Extract column edit operations to vterm_edit_col.go
Move column-level editing functions to separate file.

- vterm.go: 1320 → 1176 lines (-144)
- vterm_edit_col.go: 156 lines (new)
- InsertColumns (DECIC), DeleteColumns (DECDC)
- VT420 horizontal scrolling features
- All tests pass
- Tested with bin/texelterm - works correctly
